refactor: Model, View 폴더 분리 및 OutputView 추가
- Model, View 폴더를 생성하여 도메인 로직과 UI 로직의 관심사 분리
- 출력 책임을 전담하는 OutputView 객체 추가
- 기존 파일들을 역할에 맞는 폴더로 이동하여 디렉토리 구조 개선
