+++ src/model/Lotto.js
